Update 12DataTables.md (#254)
* Update 12DataTables.md

Added reference to HTML specification for allowed ARIA roles for a table.

* Update 12DataTables.md

Removed note stating the the test should fail if a `<table>` element also explicitly defines role="presentation". Although the specification provides guidance that aria roles SHOULD NOT duplicate the native semantics, it is stated as guidance and is non-normative.